Partially implemented autofinish
A popup menu will appear with item names that can be used to complete
the word that is currently being typed.  It can be selected using the
mouse or the arrows + return key and it will be copied to the text
field.

issues
 - if the mouse is used to choose a word, then the return key won't work
for adding unless another key is used first.
 - if the add button is used, then the arrow keys will automatically
select the first word of autofinish of the next word searched
 - if selecting a word using the arrows and return, the word will be
copied to the textfield, but the popup menu will remain on top and
visible